+++ litestar/contrib/jwt/__init__.py
@@ -10,6 +10,7 @@ from litestar.contrib.jwt.middleware import (
     JWTAuthenticationMiddleware,
     JWTCookieAuthenticationMiddleware,
 )
+from litestar.utils import warn_deprecation
 
 __all__ = (
     "BaseJWTAuth",
@@ -20,4 +21,12 @@ __all__ = (
     "OAuth2Login",
     "OAuth2PasswordBearerAuth",
     "Token",
+)
+
+warn_deprecation(
+    deprecated_name="litestar.contrib.jwt",
+    version="2.3.2",
+    kind="import",
+    removal_in="3.0",
+    info="importing from 'litestar.contrib.jwt' is deprecated, please import from 'litestar.security.jwt' instead",
 )
